Add Variables serialization with protobuf
All current type of variables can be serialized (including programs and
as a consequence actions).
We also add the ability to listen to sessions start and close from the
stack to be able to save the stack at the end of sessions and not at
each satck operation.
The goal is to use the session as the execution of a list of commands
when running in interactive mode.